+++ lib/utils/mdComponents.js
@@ -11,8 +11,11 @@ import {
   colors,
   Editorial,
   convertStyleToRem,
-  pxToRem
+  pxToRem,
+  slug
 } from '@project-r/styleguide'
+
+import Anchor from '../../components/Anchor'
 
 const styles = {
   h2: css({
@@ -57,7 +60,12 @@ const styles = {
 
 const mdComponents = {
   h1: ({ children }) => <H1 {...styles.h1}>{children}</H1>,
-  h2: H2,
+  h2: ({ children }) => (
+    <H2>
+      {typeof children === 'string' && <Anchor id={slug(children)} />}
+      {children}
+    </H2>
+  ),
   a: A,
   p: P,
   strong: ({ children }) => (
@@ -84,6 +92,7 @@ export const light = {
   ),
   h2: ({ children }) => (
     <h2 {...styles.h2} style={lightStyle}>
+      {typeof children === 'string' && <Anchor id={slug(children)} />}
       {children}
     </h2>
   ),
